Only fetch examples for aggregates returned to the user
Rather than fetch examples for all the aggregates, then sort the
aggregates and figure out which ones are returned in the response, do
it the other way around: sort the aggregates, then fetch examples for
the ones which make the cut.
