+++ lib/search/presenters/aggregate_result_presenter.rb
@@ -1,10 +1,9 @@
 module Search
   class AggregateResultPresenter
-    attr_reader :aggregates, :aggregate_examples, :search_params, :registries
+    attr_reader :aggregates, :search_params, :registries
 
-    def initialize(aggregates, aggregate_examples, search_params, registries)
+    def initialize(aggregates, search_params, registries)
       @aggregates = aggregates
-      @aggregate_examples = aggregate_examples
       @search_params = search_params
       @registries = registries
     end
@@ -29,6 +28,25 @@ module Search
       end
 
       result
+    end
+
+    # This is a class method, called on the output of
+    # AggregateResultPresenter.new(...).presented_aggregates.  This is
+    # so the examples fetched (in ../query.rb) can be derived from the
+    # presented aggregates.  The alternative is conservatively
+    # fetching any example which could possibly be needed, and
+    # throwing away unnecessary ones after-the-fact.
+    #
+    # This method mutates the 'presented_aggregates' parameter.
+    def self.merge_examples(presented_aggregates, examples)
+      presented_aggregates.each do |field, aggregate|
+        field_examples = examples[field]
+        unless field_examples.nil?
+          aggregate[:options].each do |option|
+            option[:value]["example_info"] = field_examples.fetch(option[:value]["slug"], [])
+          end
+        end
+      end
     end
 
   private
@@ -79,11 +97,6 @@ module Search
         result = { "slug" => result }
       end
 
-      field_examples = aggregate_examples[field]
-
-      unless field_examples.nil?
-        result["example_info"] = field_examples.fetch(slug, [])
-      end
       result
     end
