+++ lib/search/presenters/result_set_presenter.rb
@@ -1,20 +1,16 @@
 module Search
   # Presents a combined set of results for a GOV.UK site search
   class ResultSetPresenter
-    attr_reader :es_response, :reranked, :search_params
+    attr_reader :es_response, :presented_aggregates, :reranked, :search_params
 
     # `registries` should be a map from registry names to registries,
     # which gets passed to the ResultSetPresenter class. For example:
     #
     #     { organisations: OrganisationRegistry.new(...) }
-    #
-    # `aggregate_examples` is {field_name => {aggregate_value => {total: count, examples: [{field: value}, ...]}}}
-    # ie: a hash keyed by field name, containing hashes keyed by aggregates value with
-    # values containing example information for the value.
     def initialize(search_params:,
                    es_response:,
                    registries: {},
-                   aggregate_examples: {},
+                   presented_aggregates: {},
                    schema: nil,
                    query_payload: {},
                    reranked: false)
@@ -23,7 +19,7 @@ module Search
       @aggregates = es_response["aggregations"]
       @search_params = search_params
       @registries = registries
-      @aggregate_examples = aggregate_examples
+      @presented_aggregates = presented_aggregates
       @schema = schema
       @query_payload = query_payload
       @reranked = reranked
@@ -62,10 +58,6 @@ module Search
       es_response.dig("hits", "hits").to_a.map do |raw_result|
         ResultPresenter.new(raw_result.to_hash, @registries, @schema, search_params).present
       end
-    end
-
-    def presented_aggregates
-      AggregateResultPresenter.new(@aggregates, @aggregate_examples, @search_params, @registries).presented_aggregates
     end
   end
 end
