+++ spec/unit/result_set_presenter_spec.rb
@@ -44,98 +44,6 @@ RSpec.describe Search::ResultSetPresenter do
     }.merge(extra)
   end
 
-  def sample_aggregate_data
-    {
-      "organisations" => {
-        "filtered_aggregations" => {
-          "buckets" => [
-            { "key" => "hm-magic", "doc_count" => 7 },
-            { "key" => "hmrc", "doc_count" => 5 },
-          ],
-        },
-      },
-      "organisations_with_missing_value" => {
-        "filtered_aggregations" => {
-          "doc_count" => 8,
-        },
-      },
-    }
-  end
-
-  def sample_aggregate_data_with_policy_areas
-    {
-      "organisations" => {
-        "filtered_aggregations" => {
-          "buckets" => [
-            { "key" => "hm-magic", "doc_count" => 7 },
-            { "key" => "hmrc", "doc_count" => 5 },
-          ],
-        },
-      },
-      "organisations_with_missing_value" => {
-        "filtered_aggregations" => {
-          "doc_count" => 8,
-        },
-      },
-      "policy_areas" => {
-        "filtered_aggregations" => {
-          "buckets" => [
-            { "key" => "farming", "doc_count" => 4 },
-            { "key" => "unknown_topic", "doc_count" => 5 },
-          ],
-        },
-      },
-      "policy_areas_with_missing_value" => {
-        "filtered_aggregations" => {
-          "doc_count" => 3,
-        },
-      },
-    }
-  end
-
-  def sample_org_registry
-    {
-      "hm-magic" => {
-        "link" => "/government/departments/hm-magic",
-        "title" => "Ministry of Magic",
-      },
-      "hmrc" => {
-        "link" => "/government/departments/hmrc",
-        "title" => "HMRC",
-      },
-    }
-  end
-
-  def aggregate_response_magic
-    {
-      value: {
-        "link" => "/government/departments/hm-magic",
-        "title" => "Ministry of Magic",
-        "slug" => "hm-magic",
-      },
-      documents: 7,
-    }
-  end
-
-  def aggregate_response_hmrc
-    {
-      value: {
-        "link" => "/government/departments/hmrc",
-        "title" => "HMRC",
-        "slug" => "hmrc",
-      },
-      documents: 5,
-    }
-  end
-
-  def aggregate_params(requested, options = {})
-    {
-      requested: requested,
-      order: SearchParameterParser::DEFAULT_AGGREGATE_SORT,
-      scope: :exclude_field_filter,
-    }.merge(options)
-  end
-
   def search_presenter(options)
     org_registry = options[:org_registry]
     described_class.new(
@@ -147,7 +55,7 @@ RSpec.describe Search::ResultSetPresenter do
       ),
       es_response: sample_es_response(options.fetch(:es_response, {})),
       registries: org_registry.nil? ? {} : { organisations: org_registry },
-      aggregate_examples: options.fetch(:aggregate_examples, {}),
+      presented_aggregates: options.fetch(:presented_aggregates, {}),
       schema: options.fetch(:schema, nil),
     )
   end
@@ -175,10 +83,6 @@ RSpec.describe Search::ResultSetPresenter do
 
     it "have total of 0" do
       expect(@output[:total]).to eq(0)
-    end
-
-    it "have no aggregates" do
-      expect(@output[:aggregates]).to eq({})
     end
   end
 
@@ -307,412 +211,6 @@ RSpec.describe Search::ResultSetPresenter do
         "slug" => "farming",
       }]).to eq result["policy_areas"]
     end
-  end
-
-  context "results with aggregates" do
-    before do
-      @output = described_class.new(
-        search_params: Search::QueryParameters.new(
-          start: 0,
-          aggregates: { "organisations" => aggregate_params(1) },
-          aggregate_name: :aggregates,
-        ),
-        es_response: sample_es_response("aggregations" => sample_aggregate_data),
-      ).present
-    end
-
-    it "have aggregates" do
-      expect(@output.keys).to include(:aggregates)
-    end
-
-    it "have correct number of aggregates" do
-      expect(@output[:aggregates].length).to eq(1)
-    end
-
-    it "have correct number of aggregate values" do
-      expect(@output[:aggregates]["organisations"][:options].length).to eq(1)
-    end
-
-    it "include requested aggregate scope" do
-      aggregate = @output[:aggregates]["organisations"]
-      expect(:exclude_field_filter).to eq(aggregate[:scope])
-    end
-
-    it "have correct top aggregate value value" do
-      expect(
-        value: { "slug" => "hm-magic" },
-        documents: 7,
-      ).to eq @output[:aggregates]["organisations"][:options][0]
-    end
-
-    it "have correct number of documents with no value" do
-      expect(@output[:aggregates]["organisations"][:documents_with_no_value]).to eq(8)
-    end
-
-    it "have correct total number of options" do
-      expect(@output[:aggregates]["organisations"][:total_options]).to eq(2)
-    end
-
-    it "have correct number of missing options" do
-      expect(@output[:aggregates]["organisations"][:missing_options]).to eq(1)
-    end
-  end
-
-  context "results with aggregates and a filter applied" do
-    before do
-      @output = described_class.new(
-        search_params: Search::QueryParameters.new(
-          start: 0,
-          filters: [text_filter("organisations", %w[hmrc])],
-          aggregates: { "organisations" => aggregate_params(2) },
-          aggregate_name: :aggregates,
-        ),
-        es_response: sample_es_response("aggregations" => sample_aggregate_data),
-      ).present
-    end
-
-    it "have aggregates" do
-      expect(@output.keys).to include(:aggregates)
-    end
-
-    it "have correct number of aggregates" do
-      expect(@output[:aggregates].length).to eq(1)
-    end
-
-    it "have correct number of aggregate values" do
-      expect(@output[:aggregates]["organisations"][:options].length).to eq(2)
-    end
-
-    it "have selected aggregate first" do
-      expect(
-        value: { "slug" => "hmrc" },
-        documents: 5,
-      ).to eq @output[:aggregates]["organisations"][:options][0]
-    end
-
-    it "have unapplied aggregate value second" do
-      expect(
-        value: { "slug" => "hm-magic" },
-        documents: 7,
-      ).to eq @output[:aggregates]["organisations"][:options][1]
-    end
-
-    it "have correct number of documents with no value" do
-      expect(@output[:aggregates]["organisations"][:documents_with_no_value]).to eq(8)
-    end
-
-    it "have correct total number of options" do
-      expect(@output[:aggregates]["organisations"][:total_options]).to eq(2)
-    end
-
-    it "have correct number of missing options" do
-      expect(@output[:aggregates]["organisations"][:missing_options]).to eq(0)
-    end
-  end
-
-  context "results with aggregates and a filter which matches nothing applied" do
-    before do
-      @output = described_class.new(
-        search_params: Search::QueryParameters.new(
-          start: 0,
-          filters: [text_filter("organisations", %w[hm-cheesemakers])],
-          aggregates: { "organisations" => aggregate_params(1) },
-          aggregate_name: :aggregates,
-        ),
-        es_response: sample_es_response("aggregations" => sample_aggregate_data),
-      ).present
-    end
-
-    it "have aggregates" do
-      expect(@output.keys).to include(:aggregates)
-    end
-
-    it "have correct number of aggregates" do
-      expect(@output[:aggregates].length).to eq(1)
-    end
-
-    it "have correct number of aggregate values" do
-      expect(@output[:aggregates]["organisations"][:options].length).to eq(2)
-    end
-
-    it "have selected aggregate first" do
-      expect(
-        value: { "slug" => "hm-cheesemakers" },
-        documents: 0,
-      ).to eq @output[:aggregates]["organisations"][:options][0]
-    end
-
-    it "have unapplied aggregate value second" do
-      expect(
-        value: { "slug" => "hm-magic" },
-        documents: 7,
-      ).to eq @output[:aggregates]["organisations"][:options][1]
-    end
-
-    it "have correct number of documents with no value" do
-      expect(@output[:aggregates]["organisations"][:documents_with_no_value]).to eq(8)
-    end
-
-    it "have correct total number of options" do
-      expect(@output[:aggregates]["organisations"][:total_options]).to eq(2)
-    end
-
-    it "have correct number of missing options" do
-      expect(@output[:aggregates]["organisations"][:missing_options]).to eq(1)
-    end
-  end
-
-  context "results with aggregate counting only" do
-    before do
-      @output = described_class.new(
-        search_params: Search::QueryParameters.new(
-          start: 0,
-          aggregates: { "organisations" => aggregate_params(0) },
-          aggregate_name: :aggregates,
-        ),
-        es_response: sample_es_response("aggregations" => sample_aggregate_data),
-      ).present
-    end
-
-    it "have correct number of aggregates" do
-      expect(@output[:aggregates].length).to eq(1)
-    end
-
-    it "have no aggregate values" do
-      expect(@output[:aggregates]["organisations"][:options].length).to eq(0)
-    end
-
-    it "have correct number of documents with no value" do
-      expect(@output[:aggregates]["organisations"][:documents_with_no_value]).to eq(8)
-    end
-
-    it "have correct total number of options" do
-      expect(@output[:aggregates]["organisations"][:total_options]).to eq(2)
-    end
-
-    it "have correct number of missing options" do
-      expect(@output[:aggregates]["organisations"][:missing_options]).to eq(2)
-    end
-  end
-
-  context "results with aggregates sorted by ascending count" do
-    before do
-      org_registry = sample_org_registry
-      @output = search_presenter(
-        es_response: { "aggregations" => sample_aggregate_data },
-        aggregates: { "organisations" => aggregate_params(10, order: [[:count, 1]]) },
-        org_registry: org_registry,
-      ).present
-    end
-
-    it "have aggregates sorted by ascending count" do
-      expect([
-        aggregate_response_hmrc,
-        aggregate_response_magic,
-      ]).to eq @output[:aggregates]["organisations"][:options]
-    end
-  end
-
-  context "results with aggregates sorted by descending count" do
-    before do
-      org_registry = sample_org_registry
-      @output = search_presenter(
-        es_response: { "aggregations" => sample_aggregate_data },
-        aggregates: { "organisations" => aggregate_params(10, order: [[:count, -1]]) },
-        org_registry: org_registry,
-      ).present
-    end
-
-    it "have aggregates sorted by descending count" do
-      expect([
-        aggregate_response_magic,
-        aggregate_response_hmrc,
-      ]).to eq @output[:aggregates]["organisations"][:options]
-    end
-  end
-
-  context "results with aggregates sorted by ascending slug" do
-    before do
-      org_registry = sample_org_registry
-      @output = search_presenter(
-        es_response: { "aggregations" => sample_aggregate_data },
-        aggregates: { "organisations" => aggregate_params(10, order: [[:"value.slug", 1]]) },
-        aggregate_name: :aggregates,
-        org_registry: org_registry,
-      ).present
-    end
-
-    it "have aggregates sorted by ascending slug" do
-      expect([
-        aggregate_response_magic,
-        aggregate_response_hmrc,
-      ]).to eq @output[:aggregates]["organisations"][:options]
-    end
-  end
-
-  context "results with aggregates sorted by ascending link" do
-    before do
-      org_registry = sample_org_registry
-      @output = search_presenter(
-        es_response: { "aggregations" => sample_aggregate_data },
-        aggregates: { "organisations" => aggregate_params(10, order: [[:"value.link", 1]]) },
-        org_registry: org_registry,
-      ).present
-    end
-
-    it "have aggregates sorted by ascending link" do
-      expect([
-        aggregate_response_magic,
-        aggregate_response_hmrc,
-      ]).to eq @output[:aggregates]["organisations"][:options]
-    end
-  end
-
-  context "results with aggregates sorted by ascending title" do
-    before do
-      org_registry = sample_org_registry
-      @output = search_presenter(
-        es_response: { "aggregations" => sample_aggregate_data },
-        aggregates: { "organisations" => aggregate_params(10, order: [[:"value.title", 1]]) },
-        org_registry: org_registry,
-      ).present
-    end
-
-    it "have aggregates sorted by ascending title" do
-      expect([
-        aggregate_response_hmrc,
-        aggregate_response_magic,
-      ]).to eq @output[:aggregates]["organisations"][:options]
-    end
-  end
-
-  context "results with aggregates and an org registry" do
-    before do
-      org_registry = sample_org_registry
-
-      @output = described_class.new(
-        search_params: Search::QueryParameters.new(
-          start: 0,
-          aggregates: { "organisations" => aggregate_params(1), "policy_areas" => aggregate_params(1) },
-          aggregate_name: :aggregates,
-        ),
-        es_response: sample_es_response("aggregations" => sample_aggregate_data_with_policy_areas),
-        registries: { organisations: org_registry },
-      ).present
-    end
-
-    it "have aggregates" do
-      expect(@output.keys).to include(:aggregates)
-    end
-
-    it "have correct number of aggregates" do
-      expect(@output[:aggregates].length).to eq(2)
-    end
-
-    it "have correct number of aggregate values" do
-      expect(@output[:aggregates]["organisations"][:options].length).to eq(1)
-      expect(@output[:aggregates]["policy_areas"][:options].length).to eq(1)
-    end
-
-    it "have org aggregate value expanded" do
-      expect(
-        value: {
-          "link" => "/government/departments/hm-magic",
-          "title" => "Ministry of Magic",
-          "slug" => "hm-magic",
-        },
-        documents: 7,
-      ).to eq @output[:aggregates]["organisations"][:options][0]
-    end
-
-    it "have topic aggregate value un-expanded" do
-      expect(
-        value: { "slug" => "unknown_topic" },
-        documents: 5,
-      ).to eq @output[:aggregates]["policy_areas"][:options][0]
-    end
-
-    it "have correct number of documents with no value" do
-      expect(@output[:aggregates]["organisations"][:documents_with_no_value]).to eq(8)
-      expect(@output[:aggregates]["policy_areas"][:documents_with_no_value]).to eq(3)
-    end
-
-    it "have correct total number of options" do
-      expect(@output[:aggregates]["organisations"][:total_options]).to eq(2)
-      expect(@output[:aggregates]["policy_areas"][:total_options]).to eq(2)
-    end
-
-    it "have correct number of missing options" do
-      expect(@output[:aggregates]["organisations"][:missing_options]).to eq(1)
-      expect(@output[:aggregates]["policy_areas"][:missing_options]).to eq(1)
-    end
-  end
-
-  context "results with aggregate examples" do
-    before do
-      org_registry = sample_org_registry
-
-      @output = described_class.new(
-        search_params: Search::QueryParameters.new(
-          start: 0,
-          aggregates: { "organisations" => aggregate_params(1) },
-          aggregate_name: :aggregates,
-        ),
-        es_response: sample_es_response("aggregations" => sample_aggregate_data),
-        registries: { organisations: org_registry },
-        aggregate_examples: { "organisations" => {
-          "hm-magic" => {
-            "total" => 1,
-            "examples" => [{ "title" => "Ministry of Magic" }],
-          },
-        } },
-      ).present
-    end
-
-    it "have aggregates" do
-      expect(@output.keys).to include(:aggregates)
-    end
-
-    it "have correct number of aggregates" do
-      expect(@output[:aggregates].length).to eq(1)
-    end
-
-    it "have correct number of aggregate values" do
-      expect(@output[:aggregates]["organisations"][:options].length).to eq(1)
-    end
-
-    it "have org aggregate value expanded, and include examples" do
-      expect(
-        value: {
-          "link" => "/government/departments/hm-magic",
-          "title" => "Ministry of Magic",
-          "slug" => "hm-magic",
-          "example_info" => {
-            "total" => 1,
-            "examples" => [
-              { "title" => "Ministry of Magic" },
-            ],
-          },
-        },
-        documents: 7,
-      ).to eq @output[:aggregates]["organisations"][:options][0]
-    end
-
-    it "have correct number of documents with no value" do
-      expect(@output[:aggregates]["organisations"][:documents_with_no_value]).to eq(8)
-    end
-
-    it "have correct total number of options" do
-      expect(@output[:aggregates]["organisations"][:total_options]).to eq(2)
-    end
-
-    it "have correct number of missing options" do
-      expect(@output[:aggregates]["organisations"][:missing_options]).to eq(1)
-    end
-  end
-
-  def text_filter(field_name, values)
-    SearchParameterParser::TextFieldFilter.new(field_name, values, :filter, :any)
   end
 
   context "reranked results" do
